requirements.txt
#setup
wheel

#utils
flake8
pytest
termcolor
argparse
pathlib

#datetime
pytz

#web-scraping
requests
selenium

#parallelism
futures

#cloudstorage
google-cloud-storage
